fix: move git-ssh files to one folder
