+++ services/orchest-webserver/client/src/Routes.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { Redirect, Route, Switch, useLocation } from "react-router-dom";
+import { ConfigureGitSshView } from "./config-git-ssh-view/ConfigureGitSshView";
 import { EnvironmentsView } from "./environments-view/EnvironmentsView";
 import { JobsView } from "./jobs-view/JobsView";
 import { NotificationSettingsView } from "./notification-settings-view/NotificationSettingsView";
@@ -7,7 +8,6 @@ import PipelineView from "./pipeline-view/PipelineView";
 import { ProjectsView } from "./projects-view/ProjectsView";
 import { getOrderedRoutes, siteMap } from "./routingConfig";
 import { SettingsView } from "./settings-view/SettingsView";
-import { ConfigureGitSshView } from "./views/ConfigureGitSshView";
 import ConfigureJupyterLabView from "./views/ConfigureJupyterLabView";
 import HelpView from "./views/HelpView";
 import JupyterLabView from "./views/JupyterLabView";
